Converted to Swift 3
Added an update(option:) function to Study to facilitate manually adding and updating options. The former medthod of inserting directly into the Option set does not guarantee that the old Option will be replaced with the new Option so updates require and explicit remove and insert, which is exactly what updat(option:) does.